[add data] moved converters index file
