+++ src/plugins/kibana/server/lib/converters/ingest_pipeline_api_kibana_to_es_converter.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import * as ingestProcessorApiKibanaToEsConverters from './ingest_processor_api_kibana_to_es_converters';
+import * as ingestProcessorApiKibanaToEsConverters from '../processors/converters';
 
 export default function ingestPipelineApiKibanaToEsConverter(pipelineApiDocument) {
   return {
