+++ src/plugins/kibana/server/lib/processors/converters.js
